Moved API call for connections data and state to main component App. Multi input filter to sow street specific data ved
