+++ package.json
@@ -35,8 +35,8 @@
     "lodash": "^4.17.14",
     "morgan": "^1.9.1",
     "pg-promise": "^8.7.3",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
     "react-select": "^3.0.4",
     "webpack": "^2.2.1"
   }
